extract grid into separate component
